Improving design
